Move API call from App to Discovery module
Previously, the api call was executed inside App using an effect hook.
This resulted in a call tom the API with every render of App. This has
now been moved to the discovery module where a button below the filters
executes the Api http call instead. It also sets the returned json as
state inside App (as previously) and hydrates a session var with the
returned json. The latter is still being picked up by App to re-hydrate
state.