Refactor filter arrays in discoveryHelper
Replace arrays with objects for each column header. Each column now has
an object with name, label, and dataType keys.

The filter arrays contain an object for each database column. The object
describes the name of the data column, the label, and the dataType.
The latter is not the data type the database returns but rather the data
type the data can be coerced to. This is important for the SQL "where"
part of the query string submitted to the API. Essentially, a data
column can contain 4 types of different data which beckon different
filter methods:
 1) string data (e.g. names): filter with a text input field
 2) number data (e.g. ID or coordinates): filter with >, <, = and number
    input field
 3) enum data (e.g. type, classifications, etc): filter with enum
    drop-down box
 4) date data: filter with <, >, = date input field