Implement 'where' filter for data type Number
Any column names with data type Number can be checked which displays a
filter component that allows the user to select the comparison
operator (<, >, =) and enter a numeric value. All selected operators and
values are stateful in the discovery.js module.

The getWhereQueryString in the calTechApiRequest module appends all the
where states in a query string which is then appended to the url.