Replace API call button with automatic API call
Previously, and mostly for ease of checking, the discovery component
displayed a button that allowed the user to execute the API call after
making column selections (checkboxes) and choosing appropriate filters.
This button has been removed and replaced with an automatic API call for
any state change in those states that track the checked columns and
filters for number, string, enum, and date filters.

To achieve the automatic execution of the API call function a new module
utilityFunctions.js was created which contains a custom hook that
executes a user specified custom callback function if the function call
is not the first call (avoids calling it when the component is rendered)
and if the state of the provided array has changed.