Implement download filtered data feature
A generic dropdown button has been created (src/components/button). The
button takes a number of props, including the button label, the
queryExoplanetDatabase API request function, a function that is called
when one of the dropdown items has been clicked on, and an object for
each dropdown item.

In discovery.js, the dropdownItemClick handler is implemented. Depending
on the format chosen, a suitable API request is made, and following
a successful resolution of the API promise, the data is then passed to
the downloadData helper function.

Add commentary to App.js.