Fix download button
The switch control statement now defines the csv file type as the
default case.

Also, the bug displaying an empty data type when the download button is
clicked has been squashed. This bug was caused by providing an unused
prop to the dropdownButton component. The component itself then
deconstructs the first two props and collects the remaining props in the
dropdownItems spread. The unused prop caused the empty button inside the
dropdown list.