Replace auto-API-request with Update button
Create simple button component which requires two props, namely the
button name and either a url string which the button will redirect the
user to or a function which is being executed upon clicking the button.

Note: the redirect feature has not yet been implemented

The queryExoplanetDatabase function is not passed down to the filter
components anymore as updates now need to be effected by clicking the
button.