Session-store selected columns and filters
Previously, the data response from the API request was stored in
session storage. This did not work well since there were configurations
for which the returned data was too large for browser's session storage
to handle / save it. Also, retrieving the data from the API is cheap and
fast but users spend most time with fine-tuning the column selection
and the filter options.

As a result, the response data from the API is not stored in session
storage anymore. Instead, the application now stores the selectedColumns
and whereFilter objects in session storage.

The simpleButton component has also been amended to such effect that a
single click on the button does not execute the provided callback more
than once. This was achieved by removing the buttonEffectAsync callback
from the dependency list inside the useEffect hook.