Debug axios factory function
Previously the factory function axios.create({baseURL}) was used to set
the baseURL where the returned function would then be called with an
empty url string and the params object. For some unknown reason this
approach has stopped working as axios throws a "config url" error when
an empty url string is provided to the returned function object.
Instead, the new factory function axios.create() takes no arguments,
with the full CalTech url string provided to the returned function
instead.