Debug refresh error in Glossary view
Hitting the refresh button in the browser when on the Glossary page
caused the following error "can't access property map, glossaryArray is
undefined". This happened because glossaryArray is assigned inside a
useEffect in the App component after the first render is done. That
means that for the first render glossaryArray is not defined.

I have added a second condition to line 32 in glossaryList.js that
controls for this.