Add data table and infinite scroll
Add simple data table implementation that displays an un-styled
tabular representation of the received CalTech data. Since the data can
contain up to 10,000 rows, an infinite scroll has been implemented to
avoid performance issues associated with loading 10,000 rows.
Every time the user scrolls to the bottom of the data table, an
additional row is added to the table.

Next stage:
- experiment with divs and drag&drop of columns
- quite likely that a table set up may be too restrictive