Implement Levishtein distance as search algo
If the search term is contained inside the target string, the
Levenshtein distance will be exactly the delta between the length of the
target string and the search term. If the Levenshtein distance is bigger
than the delta between both lengths then that means that additional
characters had to be either added, swapped, or deleted. Hence, if the
Levenshtein distance + fuzzyChars is smaller-equal the difference in
length between the target string and word then return true.

e.g: word="long"  target="longer" => Levenshtein distance=2 => allowing
for one fuzzy character implies that a Levishtein distance of 3 should
still return "longer" as a search result even if the user has typed in
"lone" instead of "long".