Debug processing of koi_quarters API response
New function is used to pre-process unparsed data received from API
if koi_quarters data is received. That is necessary because any response
for that data is sent in string format and is not automatically unparsed
unlike the responses to any other data column.

The new pre-processor function is only relevant for the data set
associated with "koi_quarters" which seems to be the only data set
that returns string data, containing 4 byte binary string that is not
wrapped in apostrophes. The data is in the form of
"[\n{\"koi_quarters\":01111111111111111000000000000000}\n]\n".
The binary cannot be parsed unless it is wrapped in apostrophes.
This function returns the data with all binary data wrapped in
apostrophes.